Clean code function name

+++ MongoGoogle/Controller/AuthenticationController.go
@@ -77,9 +77,7 @@ func Mark1() {
 		if tokenString != "" {
 			res.Header().Set("Content-Type", "application/json")
 			json.NewEncoder(res).Encode(tokenString)
-		} else {
 		}
-
 	})
 
 	//Our service that serves login functionality
@@ -96,7 +94,6 @@ func Mark1() {
 		}
 		authHeader := req.Header.Get("Authorization")
 		tokenService.TokenAppLoginLogic(res, req, authHeader, requestBody.Email, requestBody.Password)
-
 	})
 
 	//Our service that serves registration functionality
@@ -127,27 +124,19 @@ func Mark1() {
 			http.Error(res, "Error extracting user from token", http.StatusInternalServerError)
 			return
 		}
-		fmt.Println("Logout" + " " + user.Email)
-
 		tokenExpString, _ := tokenService.GenerateToken(user, time.Second)
 
 		res.Header().Set("Content-Type", "application/json")
-		json.NewEncoder(res).Encode(struct {
-			Token string `json:"token"`
-		}{
-			Token: user.Email + "Successfully logged out, this is your new bearer token: \n" + tokenExpString,
-		})
+		json.NewEncoder(res).Encode("Token:" + user.Email + "Successfully logged out, this is your new bearer token: " + tokenExpString)
 	})
 
 	// Verifies the user with the specified email address.
 	r.HandleFunc("/verify/{email}", func(res http.ResponseWriter, req *http.Request) { //postman
-
 		vars := mux.Vars(req)
 		email := vars["email"]
 		if db.VerifyUser(email) {
-			fmt.Fprintf(res, email)
+			fmt.Println(res, email)
 		}
-
 	})
 
 	// Registers a new company using the provided email address for authentication.
@@ -173,7 +162,7 @@ func Mark1() {
 				return
 			}
 
-			if db.ValidComapnyName(companyData.Name) {
+			if db.FindComapnyName(companyData.Name) {
 				fmt.Printf("Company exist\n")
 			} else {
 				db.SetUserRole(user.ID, "Owner")
@@ -182,7 +171,7 @@ func Mark1() {
 				user, _ := db.GetUserData(user.Email)
 				token, _ := tokenService.GenerateToken(user, time.Hour)
 				res.Header().Set("Content-Type", "application/json")
-				json.NewEncoder(res).Encode("You are successfully created new company: " + companyData.Name + ", this is your new bearer token\n" + token)
+				json.NewEncoder(res).Encode("You are successfully created new company: " + companyData.Name + ", this is your new bearer token: " + token)
 			}
 		} else {
 			res.Header().Set("Content-Type", "application/json")
@@ -209,20 +198,28 @@ func Mark1() {
 			http.Error(res, "Error decoding request body", http.StatusBadRequest)
 			return
 		}
-		if tokenUser != nil && tokenUser.Valid {
-			if requestBody.CompanyName == "" {
-				http.Error(res, "Company name is required", http.StatusBadRequest)
+		if requestBody.CompanyName == "" {
+			http.Error(res, "Company name is required", http.StatusBadRequest)
+			return
+		} else {
+			company, err := db.FindCompanyByName(requestBody.CompanyName, res)
+			if err == false {
+				res.Header().Set("Content-Type", "application/json")
+				json.NewEncoder(res).Encode("You don't have any company")
 				return
 			}
-			db.SetUserRole(user.ID, "User")
-			db.DeleteCompany(requestBody.CompanyName)
-			user, _ := db.GetUserData(user.Email)
-			token, _ := tokenService.GenerateToken(user, time.Hour)
-			res.Header().Set("Content-Type", "application/json")
-			json.NewEncoder(res).Encode("You are successfully deleted company " + requestBody.CompanyName + ", this is your new bearer token\n" + token)
-		} else {
-			res.Header().Set("Content-Type", "application/json")
-			json.NewEncoder(res).Encode("You are not owner of" + requestBody.CompanyName)
+			if tokenUser != nil && tokenUser.Valid && user.ID == company.Owner {
+				db.SetUserRole(user.ID, "User")
+				db.DeleteCompany(requestBody.CompanyName)
+				user, _ := db.GetUserData(user.Email)
+				token, _ := tokenService.GenerateToken(user, time.Hour)
+				res.Header().Set("Content-Type", "application/json")
+				json.NewEncoder(res).Encode("You are successfully deleted company " + requestBody.CompanyName + ", this is your new bearer token: " + token)
+			} else {
+				res.Header().Set("Content-Type", "application/json")
+				json.NewEncoder(res).Encode("You are not owner of" + requestBody.CompanyName)
+			}
+
 		}
 	})
